Migrated covid19.py
